High or Low(Guessing )Game
Initially, the system randomly selects a number. Based on the user's guess, the system tells if the user's guess is lower or higher than the system's number. There are two levels: hard and easy. Depending on the level of difficulty, the user gets the number of attempts to guess the number.